Move edugain cronjob to post_tasks
We can only run the edugain cronjob when the whole node is setup, including
running all the handlers. This is easiest in a post_tasks block, hence we add a
edugain_cronjob post task.
The include path is the playbook directory only, so we just include it
specifying its directory.