Set NameIDPolicy to false instead of null
The NameIDPolicy variable determines which type of NameIDPolicy is requested in
the AuthnRequest (see https://simplesamlphp.org/docs/stable/saml:sp). Leaving it
unset or setting it to null is equivalent to requesting 'transient'. Since we
never use a transient NameID, and since transient can break using ADFS IdPs, it
is better to just set it to false which means not putting any NameIDPolicy in
the request.